+++ CHANGELOG.md
@@ -3,6 +3,8 @@
   * solve many history issues
   * add new tags `Elektro Komponenten` & `Brickbuild Bäume`, change `Gleise` to `Gleise und Deko`, thx @jgrajer
   * fix some tag issues
+  * add fetch interval for changes
+  * new welcome page with support
 * [27.01.2022]
   * add time ago into product history
   * active product in url